Extract hardware abstraction to util folder
